% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apiTrusts}
\alias{apiTrusts}
\title{NHS trust admissions data from the legacy COVID API from 2021}
\format{
A 72928 line data frame:
\describe{
  \item{date}{the ONS code for the ares}
  \item{area_name}{The name for the area}
  \item{area_type}{the type of area (LSOA, DZ or LGD)}
  \item{area_code}{the NHS ODS code}
  \item{metric}{hospital cases}
  \item{metric_name}{hospital cases}
  \item{value}{the count}
}
}
\usage{
apiTrusts
}
\description{
NHS trust admissions data from the legacy COVID API from 2021
}
\concept{datasets}
\keyword{datasets}
