% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{arcgisServiceInfo}
\alias{arcgisServiceInfo}
\title{Get feature service information from an argis rest service}
\usage{
arcgisServiceInfo(url, ...)
}
\arguments{
\item{url}{the base rest service url}

\item{...}{
  Named arguments passed on to \code{\link[=.cache_download]{.cache_download}}\describe{
    \item{\code{.nocache}}{if set to TRUE all caching is disabled}
\item{\code{.cache}}{the location of the downloaded files}
\item{\code{.stale}}{how long to leave this file before replacing it.}
\item{\code{.extn}}{the file name extension}
}}
}
\value{
a dataframe of available feature and map services
}
\description{
Get feature service information from an argis rest service
}
\examples{
url = "https://services1.arcgis.com/ESMARspQHYMw9BZ9/ArcGIS/rest/services"
arcgisServiceInfo(url)
}
