% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{CA19}
\alias{CA19}
\title{Map data: CA19}
\format{
An object of class \code{character} of length 1.
}
\usage{
CA19
}
\description{
Map data: CA19
}
\references{
\url{https://www.spatialdata.gov.scot/geonetwork/srv/eng/catalog.search#/metadata/1cd57ea6-8d6e-412b-a9dd-d1c89a80ad62}
}
\keyword{datasets}
