% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_clear}
\alias{cache_clear}
\title{Clear data from the passthrough cache for complex or long running operations}
\usage{
cache_clear(
  cache = getOption("arear.cache.dir", default = rappdirs::user_cache_dir("arear")),
  interactive = TRUE
)
}
\arguments{
\item{cache}{the location of the cache as a directory. May get its value from options("arear.cache.dir") or the default value of rappdirs::user_cache_dir("arear")}

\item{interactive}{if FALSE will delete the files without warning. Defaults to TRUE}
}
\value{
nothing
}
\description{
Clear data from the passthrough cache for complex or long running operations
}
