% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{CCG20}
\alias{CCG20}
\title{Map data: CCG20}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 135 rows and 6 columns.
}
\usage{
CCG20
}
\description{
Map data: CCG20
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/clinical-commissioning-groups-april-2020-generalised-clipped-boundaries-en}
}
\keyword{datasets}
