% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{createNeighbourNetwork}
\alias{createNeighbourNetwork}
\title{create a neighbourhood network from a shapefile}
\usage{
createNeighbourNetwork(shape, idVar = "code", bridges = tibble::tibble(), ...)
}
\arguments{
\item{shape}{- a sf object, if not present will be loaded from cache}

\item{idVar}{- the varable containing the coded identifier of the map}

\item{bridges}{-}
}
\value{
an edge list of ids with from and to columns
}
\description{
create a neighbourhood network from a shapefile
}
