% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{CTRY19}
\alias{CTRY19}
\title{Map data: CTRY19}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 6 columns.
}
\usage{
CTRY19
}
\description{
Map data: CTRY19
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/countries-december-2019-boundaries-uk-bgc/data}
}
\keyword{datasets}
