% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{CTYUA19}
\alias{CTYUA19}
\title{Map data: CTYUA19}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 216 rows and 6 columns.
}
\usage{
CTYUA19
}
\description{
Map data: CTYUA19
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/counties-and-unitary-authorities-april-2019-boundaries-ew-buc-1}
}
\keyword{datasets}
