% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{devonlocality}
\alias{devonlocality}
\title{An administrative classification of hospital catchment areas by postcode in the Devon region}
\format{A 52157 line data frame:
\describe{
  \item{pcds}{The postcode of an area}
  \item{locality}{the administrative NHS Devon locality}
  \item{trustId}{the id of the main NHS trust in the locality}
  \item{trustame}{the name of the main NHS trust in the locality}
  \item{code}{the LSOA11 geographical code}
  \item{codeType}{the type of area - always LSOA11}
  \item{altCode}{NA}
}}
\source{
\describe{
  \item{Locality Postcodes.xlsx}{personal communication}
  \item{Post codes to LSOA11 codes, ONSPD}{https://www.ons.gov.uk/methodology/geography/geographicalproducts/postcodeproducts}
}
}
\usage{
devonlocality
}
\description{
An administrative classification of hospital catchment areas by postcode in the Devon region
}
\keyword{datasets}
