% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-maps.R
\name{downloadGeojson}
\alias{downloadGeojson}
\title{Download a geojson url, standardise it and cache the result}
\usage{
downloadGeojson(
  geojsonUrl,
  codeCol = "code",
  nameCol = "name",
  altCodeCol = NULL,
  codeType = NA_character_,
  simplify = FALSE,
  id,
  license = "unknown",
  ...
)
}
\arguments{
\item{geojsonUrl}{the URL of the geojson resource}

\item{codeCol}{- the name of the column containing the id or code}

\item{nameCol}{- the name of the column containing the label (optional - defaults to the same as codeCol)}

\item{altCodeCol}{- an optional column name containing another code type}

\item{codeType}{- the "type" of the code - optional. defaults to NA}

\item{simplify}{- do you want to simplify the map}

\item{id}{- an id for the map that can be used to retrieve it in the future (through getMap()).}

\item{license}{- an optional license string}

\item{...}{- passed to .cache, param nocache=TRUE to disable caching}
}
\value{
the sf object for this geoJson
}
\description{
Download a geojson url, standardise it and cache the result
}
