% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{DZ11}
\alias{DZ11}
\title{Map data: DZ11}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 6976 rows and 6 columns.
}
\usage{
DZ11
}
\description{
Map data: DZ11
}
\references{
\url{https://data.gov.uk/dataset/ab9f1f20-3b7f-4efa-9bd2-239acf63b540/data-zone-boundaries-2011}
}
\keyword{datasets}
