% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{featureServerInfo}
\alias{featureServerInfo}
\title{Get layer information from a feature service}
\usage{
featureServerInfo(url, ...)
}
\arguments{
\item{url}{the feature service url}

\item{...}{
  Named arguments passed on to \code{\link[=.cache_download]{.cache_download}}\describe{
    \item{\code{.nocache}}{if set to TRUE all caching is disabled}
\item{\code{.cache}}{the location of the downloaded files}
\item{\code{.stale}}{how long to leave this file before replacing it.}
\item{\code{.extn}}{the file name extension}
}}
}
\value{
a dataframe of layers
}
\description{
Get layer information from a feature service
}
\examples{
svc = "https://services1.arcgis.com/ESMARspQHYMw9BZ9/ArcGIS/rest/services"
fs = sprintf("\%s/\%s",svc,"Countries_December_2024_Boundaries_UK_BUC/FeatureServer")
url = sprintf("\%s/\%s",fs,"0")
featureServerInfo(fs)
}
