% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{featureServerLayerInfo}
\alias{featureServerLayerInfo}
\title{Get field information from a feature service layer.}
\usage{
featureServerLayerInfo(url, ...)
}
\arguments{
\item{url}{the feature service layer url. If no layer is specified the first
one is selected.}

\item{...}{
  Named arguments passed on to \code{\link[=.cache_download]{.cache_download}}\describe{
    \item{\code{.nocache}}{if set to TRUE all caching is disabled}
\item{\code{.cache}}{the location of the downloaded files}
\item{\code{.stale}}{how long to leave this file before replacing it.}
\item{\code{.extn}}{the file name extension}
}}
}
\value{
a dataframe of field names for the selected layer
}
\description{
Get field information from a feature service layer.
}
\examples{
svc = "https://services1.arcgis.com/ESMARspQHYMw9BZ9/ArcGIS/rest/services"
fs = sprintf("\%s/\%s",svc,"Countries_December_2024_Boundaries_UK_BUC/FeatureServer")
layer = sprintf("\%s/\%s",fs,"0")
featureServerLayerInfo(layer)
}
\concept{arcgis}
