% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{GBR_ISO3166_2}
\alias{GBR_ISO3166_2}
\title{Map data: GBR_ISO3166_2}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 183 rows and 6 columns.
}
\usage{
GBR_ISO3166_2
}
\description{
Map data: GBR_ISO3166_2
}
\references{
\url{https://gadm.org/download_country_v3.html}
}
\keyword{datasets}
