% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{getContainedIn}
\alias{getContainedIn}
\title{finds shapes that contain the give shape}
\usage{
getContainedIn(inputSf, outputShape, inputIdVar = "code", outputIdVar = "code")
}
\description{
finds shapes that contain the give shape
}
