% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-maps.R
\name{getMap}
\alias{getMap}
\title{Gets maps for which the metadata is known}
\usage{
getMap(
  mapId,
  sources = arear::mapsources,
  codeType = mapId,
  wd = getOption("arear.cache.dir", default = tempdir())
)
}
\arguments{
\item{mapId}{- a name of a map}

\item{sources}{- a list of map sources - see arear::mapsources}

\item{codeType}{- defaults to mapId, the codeType of the map}

\item{wd}{- optional working directory}
}
\value{
a standard sf map
}
\description{
Gets maps for which the metadata is known
}
\examples{
\dontrun{
map = getMap("NHSER20")
}
}
