% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{GOOGLE_MOBILITY}
\alias{GOOGLE_MOBILITY}
\title{Map data: GOOGLE_MOBILITY}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 151 rows and 6 columns.
}
\usage{
GOOGLE_MOBILITY
}
\description{
Map data: GOOGLE_MOBILITY
}
\references{
\url{https://github.com/datasciencecampus/google-mobility-reports-data/tree/master/geography}
}
\keyword{datasets}
