% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{HB19}
\alias{HB19}
\title{Map data: HB19}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 14 rows and 6 columns.
}
\usage{
HB19
}
\description{
Map data: HB19
}
\references{
\url{https://www.spatialdata.gov.scot/geonetwork/srv/api/records/f12c3826-4b4b-40e6-bf4f-77b9ed01dc14}
}
\keyword{datasets}
