% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{interpolateByArea}
\alias{interpolateByArea}
\title{interpolate a variable from one set of shapes to another}
\usage{
interpolateByArea(
  inputDf,
  inputShape,
  inputIdVar = "code",
  interpolateVar,
  outputShape,
  outputVars = outputShape \%>\% dplyr::groups(),
  aggregateFn = sum
)
}
\arguments{
\item{inputDf}{- a grouped dataframe containing the statistic to be interpolated}

\item{inputShape}{- an input map,}

\item{inputIdVar}{- an id shared between the grouped data fram and the input map,}

\item{interpolateVar}{- the statistic,}

\item{outputShape}{- an output map which must be grouped by the desired output,}
}
\value{
a dataframe containing the grouping columns, the outputIdVar and the interpolated value of interpolateVar
}
\description{
interpolate a variable from one set of shapes to another
}
