% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{LAD19}
\alias{LAD19}
\title{Map data: LAD19}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 382 rows and 6 columns.
}
\usage{
LAD19
}
\description{
Map data: LAD19
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/local-authority-districts-december-2019-boundaries-uk-buc}
}
\keyword{datasets}
