% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{LAD20}
\alias{LAD20}
\title{Map data: LAD20}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 379 rows and 6 columns.
}
\usage{
LAD20
}
\description{
Map data: LAD20
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/local-authority-districts-may-2020-boundaries-uk-buc}
}
\keyword{datasets}
