% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{LGD12}
\alias{LGD12}
\title{Map data: LGD12}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 11 rows and 6 columns.
}
\usage{
LGD12
}
\description{
Map data: LGD12
}
\references{
\url{https://data.gov.uk/dataset/05f72866-b72b-476a-b6f3-57bd4a768674/osni-open-data-largescale-boundaries-local-government-districts-2012}
}
\keyword{datasets}
