% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{LHB19}
\alias{LHB19}
\title{Map data: LHB19}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 6 columns.
}
\usage{
LHB19
}
\description{
Map data: LHB19
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/local-health-boards-april-2019-boundaries-wa-buc}
}
\keyword{datasets}
