% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-maps.R
\name{listStandardMaps}
\alias{listStandardMaps}
\title{List the standard maps available to download}
\usage{
listStandardMaps()
}
\value{
a vector of map names
}
\description{
List the standard maps available to download
}
