% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{londonShape}
\alias{londonShape}
\title{An outline of the London NHSER region for use as a mask}
\format{
A sf geometry with 1 row:
\describe{
  \item{codeType}{NHSER}
  \item{code}{E4000003 - the NHSER code for London}
  \item{name}{London}
  \item{altCode}{NA}
  \item{area}{the area}
  \item{geometry}{the outline}
}
}
\usage{
londonShape
}
\description{
Data available under open government licence: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
}
\keyword{datasets}
