% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{LSOA11}
\alias{LSOA11}
\title{Map data: LSOA11}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34753 rows and 6 columns.
}
\usage{
LSOA11
}
\description{
Map data: LSOA11
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/lower-layer-super-output-areas-december-2011-boundaries-ew-bgc}
}
\keyword{datasets}
