% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapsources}
\alias{mapsources}
\title{Locations of UK and international shapefiles relevant to COVID}
\format{
A list with:
\describe{
  \item{source}{the human readable location of the map}
  \item{url}{the web location of the downloadable map shapefile}
  \item{mapName}{the name of the map contained in the shapefile (which can contain multiple maps)}
  \item{codeCol}{the name of the shapefile column containing the code of the area}
  \item{nameCol}{the name of the shapefile column containing the name of the area}
  \item{altCodeCol}{the name of the shapefile column containing the an alternative code for the area}
  \item{simplify}{should the map be simplified when loaded?}
  \item{license}{license terms}
}
}
\usage{
mapsources
}
\description{
A list of URLs to get maps, and metadata about the maps in the shapefiles and
the column labelling.
}
\concept{datasets}
\keyword{datasets}
