% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{mapTheme}
\alias{mapTheme}
\title{A map theme}
\usage{
mapTheme(baseSize = 8)
}
\arguments{
\item{baseSize}{the basic font size in points.}
}
\description{
A map theme
}
\examples{
ggplot2::ggplot()+mapTheme()
}
