% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{MSOA11}
\alias{MSOA11}
\title{Map data: MSOA11}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7201 rows and 6 columns.
}
\usage{
MSOA11
}
\description{
Map data: MSOA11
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/middle-layer-super-output-areas-december-2011-boundaries-generalised-clipped-bgc-ew-v3}
}
\keyword{datasets}
