% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{NHSER20}
\alias{NHSER20}
\title{Map data: NHSER20}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 6 columns.
}
\usage{
NHSER20
}
\description{
Map data: NHSER20
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/nhs-england-regions-april-2020-boundaries-en-bgc}
}
\keyword{datasets}
