% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{OUTCODE}
\alias{OUTCODE}
\title{Map data: OUTCODE}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2880 rows and 6 columns.
}
\usage{
OUTCODE
}
\description{
Map data: OUTCODE
}
\references{
\url{https://www.opendoorlogistics.com/downloads/}
}
\keyword{datasets}
