% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{PHEC16}
\alias{PHEC16}
\title{Map data: PHEC16}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 6 columns.
}
\usage{
PHEC16
}
\description{
Map data: PHEC16
}
\references{
\url{https://ons-inspire.esriuk.com/arcgis/rest/services/Health_Boundaries/Public_Health_England_Centres_December_2016_Boundaries}
}
\keyword{datasets}
