% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{popoutArea}
\alias{popoutArea}
\title{Create a magnified pop-out panel of a map}
\usage{
popoutArea(
  shape,
  popoutShape = arear::londonShape,
  popoutPosition = c("NE", "NW", "SE", "SW"),
  popoutScale = 3,
  nudgeX = 0.25,
  nudgeY = 0.25
)
}
\arguments{
\item{shape}{The original shape}

\item{popoutShape}{The mask shape. The outer boundary of this shape will be
used as a mask to select the original shape}

\item{popoutPosition}{Which corner to place the popout NE,NW,SE or SW}

\item{popoutScale}{A factor to grow the popout area by. This is linear scale
so the popout will appear the square of this factor bigger.}

\item{nudgeX}{shift the popout panel by a small amount (in coordinate units)}

\item{nudgeY}{shift the popout panel by a small amount (in coordinate units)}
}
\value{
A new map with the content intersecting the popout area duplicated,
  expanded and placed in the specified corner.
}
\description{
Create a new map with a pop-out panel to show areas where there is a high
density of people for example. Defaults work well for London on an England
only map.
}
\concept{vis}
