% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{preview}
\alias{preview}
\title{Preview a map with POI using leaflet}
\usage{
preview(
  shape,
  shapeLabelGlue = "{name}",
  shapePopupGlue = "{code}",
  poi = NULL,
  poiLabelGlue = "{name}",
  poiPopupGlue = "{code}"
)
}
\arguments{
\item{shape}{- the map}

\item{shapeLabelGlue}{- a glue specification for the label for each shape}

\item{shapePopupGlue}{- a glue specification for the popup for each shape}

\item{poi}{- a list of points of interest as a sf object}

\item{poiLabelGlue}{- a glue specification for the label for each poi}

\item{poiPopupGlue}{- a glue specification for the popup for each poi}
}
\value{
htmlwidget
}
\description{
Preview a map with POI using leaflet
}
\concept{vis}
