% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-maps.R
\name{saveShapefile}
\alias{saveShapefile}
\title{save a shapefile to disk in the current working directory}
\usage{
saveShapefile(shape, mapId, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{shape}{- the sf shape}

\item{mapId}{- a mapId - will become the zip filename, and the filename of the zipped .shp file}

\item{dir}{- the directory (defaults to current working directory)}

\item{overwrite}{- the save function will not write over existing files unless this is set to true}
}
\value{
a the filename of the zipped shapefile
}
\description{
save a shapefile to disk in the current working directory
}
