% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{uk2019demographics}
\alias{uk2019demographics}
\title{UK small area single year of age population estimates from 2019}
\usage{
uk2019demographics()
}
\value{
a dataframe with age, gender, codeType, code, name and count
}
\description{
These estimates are appropriate for the majority of the pandemic, and are
the highest geographical resolution estimates by single year of age that
I could find.
}
\concept{datasets}
