% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{uk2019demographicsmap}
\alias{uk2019demographicsmap}
\title{Shape file related to the mid year 2019 small area estimates for England, Wales, Scotland, and Northern Ireland}
\source{
\describe{
  \item{England and Wales}{https://geoportal.statistics.gov.uk/datasets/lower-layer-super-output-areas-december-2011-boundaries-ew-bgc}
  \item{Scotland}{https://data.gov.uk/dataset/ab9f1f20-3b7f-4efa-9bd2-239acf63b540/data-zone-boundaries-2011}
  \item{Northern Ireland}{https://data.gov.uk/dataset/05f72866-b72b-476a-b6f3-57bd4a768674/osni-open-data-largescale-boundaries-local-government-districts-2012}
}
}
\usage{
uk2019demographicsmap()
}
\value{
A 41730 line data frame:
\describe{
  \item{code}{the ONS code for the ares}
  \item{name}{The name for the area}
  \item{codeType}{the type of area (LSOA, DZ or LGD)}
  \item{altCode}{NA}
}
}
\description{
There are 10 regions (mostly in Scotland) where the demographics estimates don't align with this map. This is
a small number of people
}
\details{
Data available under open government licence: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
}
