% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{uk2019retiredpopulation}
\alias{uk2019retiredpopulation}
\title{Mid year 2019 small area estimates for adult population of England, Wales, Scotland, and Northern Ireland}
\format{A 41740 line data frame:
\describe{
  \item{code}{the ONS code for the ares}
  \item{name}{The name for the area}
  \item{codeType}{the type of area (LSOA, DZ or LGD)}
  \item{population}{the size of the population}
}}
\source{
\describe{
  \item{England and Wales}{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/lowersuperoutputareamidyearpopulationestimates}
  \item{Scotland}{https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/population/population-estimates/2011-based-special-area-population-estimates/small-area-population-estimates/time-series#2018}
  \item{Northern Ireland}{https://www.opendatani.gov.uk/dataset/3333626e-b96e-4b90-82fb-474c6c03b868/resource/64bd8dc4-935f-4bdd-9232-90ff33f24732/}
}
}
\usage{
uk2019retiredpopulation
}
\description{
Small area single digit estimates are aggregated to include only over 65s, and combining gender
Data available under open government licence: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
}
\keyword{datasets}
