% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ukconnections}
\alias{ukconnections}
\title{Locations of UK bridge or ferry route start and end points}
\format{A dataframe with:
\describe{
  \item{name}{the connection}
  \item{start.lat}{the latitude of one end}
  \item{start.long}{the longditude of one end}
  \item{end.lat}{the latitude of the other end}
  \item{end.long}{the longditude of the other end}
}}
\usage{
ukconnections
}
\description{
geographical location of bridges / ferries UK connections.
picked from https://developers.google.com/maps/documentation/javascript/examples/event-click-latlng#maps_event_click_latlng-html
}
\keyword{datasets}
