% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ukcovidmap}
\alias{ukcovidmap}
\title{Shape file related to the detailed level for PHE coronavirus statistical reporting in England, Wales, Scotland, and Northern Ireland.}
\format{A 382 line data frame:
\describe{
  \item{code}{the ONS code for the ares}
  \item{name}{The name for the area}
  \item{codeType}{the type of area (LAD19)}
  \item{altCode}{NA}
}}
\source{
\describe{
  \item{England, Wales, Scotland & NI - LAD19}{https://geoportal.statistics.gov.uk/datasets/local-authority-districts-december-2019-boundaries-uk-buc}
}
}
\usage{
ukcovidmap
}
\description{
This map matches the data reported by the PHE coronavirus api when it is downloading lower tier local authority regions e.g. via
https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&metric=newCasesBySpecimenDate&format=csv
There are 2 regions here for which no data is reported in this API - the city of London & the isles of scilly.
}
\details{
Data available under open government licence: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
}
\keyword{datasets}
