% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{uklegacycovidmap}
\alias{uklegacycovidmap}
\title{Shape file related to the detailed level for PHE coronavirus statistical reporting in England, Wales, Scotland, and Northern Ireland.}
\format{A 335 line data frame:
\describe{
  \item{code}{the ONS code for the area}
  \item{name}{The name for the area}
  \item{codeType}{the type of area (LAD19, LHB, HB, LGD)}
  \item{altCode}{NA}
}}
\source{
\describe{
  \item{England & Wales - LAD19}{https://geoportal.statistics.gov.uk/datasets/local-authority-districts-december-2019-boundaries-uk-buc}
  \item{Scotland - CA19}{https://www.spatialdata.gov.scot/geonetwork/srv/eng/catalog.search#/metadata/1cd57ea6-8d6e-412b-a9dd-d1c89a80ad62}
  \item{Northern Ireland  -LGD12}{https://data.gov.uk/dataset/05f72866-b72b-476a-b6f3-57bd4a768674/osni-open-data-largescale-boundaries-local-government-districts-2012}
}
}
\usage{
uklegacycovidmap
}
\description{
This level is the level at which detailed data was reported in the first wave and was aggregated by
https://github.com/tomwhite/covid-19-uk-data
}
\details{
Data available under open government licence: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
}
\keyword{datasets}
