% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{WD11}
\alias{WD11}
\title{Map data: WD11}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8588 rows and 6 columns.
}
\usage{
WD11
}
\description{
Map data: WD11
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/wards-december-2011-boundaries-ew-bgc}
}
\keyword{datasets}
