% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{WD19}
\alias{WD19}
\title{Map data: WD19}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8887 rows and 6 columns.
}
\usage{
WD19
}
\description{
Map data: WD19
}
\references{
\url{https://geoportal.statistics.gov.uk/datasets/wards-december-2019-boundaries-ew-bgc}
}
\keyword{datasets}
