% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear-package.R
\docType{package}
\name{arear-package}
\alias{arear}
\alias{arear-package}
\title{arear: Arear: Supply demand catchment area}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Robert Challen \email{rc538@exeter.ac.uk} (\href{https://orcid.org/0000-0002-5504-7768}{ORCID})

}
\keyword{internal}
