% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{createNeighbourNetwork}
\alias{createNeighbourNetwork}
\title{create a neighbourhood network from a shapefile}
\usage{
createNeighbourNetwork(
  shape,
  idVar = "code",
  bridges = arear::ukconnections,
  queen = FALSE,
  ...
)
}
\arguments{
\item{shape}{- a sf object, if not present will be loaded from cache}

\item{idVar}{- the column containing the coded identifier of the map}

\item{bridges}{- a df with the following columns: name start.lat start.long end.lat end.long defining connections between non touching shapes (e.g. bridges / ferries / etc.)}

\item{queen}{- include neighbouring areas that touch at corners}

\item{...}{- passed on to .cached() (cache control)}
}
\value{
an edge list of ids with from and to columns
}
\description{
create a neighbourhood network from a shapefile
}
