% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{createNeighbourNetwork}
\alias{createNeighbourNetwork}
\title{create a neighbourhood network from touching regions in a shapefile, with additional capability to connect non touching areas where there may be bridges etc.}
\usage{
createNeighbourNetwork(
  shape,
  idVar = "code",
  bridges = arear::ukconnections,
  queen = FALSE,
  ...
)
}
\arguments{
\item{shape}{- a sf object, if not present will be loaded from cache}

\item{idVar}{- the column containing the coded identifier of the map}

\item{bridges}{- a df with the following columns: name start.lat start.long
end.lat end.long defining connections between non touching shapes (e.g.
bridges / ferries / etc.)}

\item{queen}{- include neighbouring areas that only touch at corners,
defaults to false.}

\item{...}{
  Named arguments passed on to \code{\link[=.cached]{.cached}}\describe{
    \item{\code{.nocache}}{an option to defeat the caching which can be set globally as options("cache.disable"=TRUE)}
}}
}
\value{
an edge list of ids with from and to columns
}
\description{
create a neighbourhood network from touching regions in a shapefile, with additional capability to connect non touching areas where there may be bridges etc.
}
\concept{analysis}
