% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-maps.R
\name{downloadMap}
\alias{downloadMap}
\title{Download a map, unpack it, and rename columns}
\usage{
downloadMap(
  zipUrl,
  mapName = NULL,
  codeCol = "code",
  nameCol = "name",
  altCodeCol = NULL,
  codeType = NA_character_,
  simplify = FALSE,
  wd = tempdir(),
  id = NULL
)
}
\arguments{
\item{zipUrl}{- the URL of the zipped shapefile}

\item{mapName}{- the layer name or map name - this is the "xyz" of a zip file containing "xyz.shp"}

\item{codeCol}{- the name of the column containing the id or code}

\item{nameCol}{- the name of the column containing the label (optional - defaults to the same as codeCol)}

\item{altCodeCol}{- an optional column name containing another code type}

\item{codeType}{- the "type" of the code - optional. defaults to NA}

\item{simplify}{- do you want to implify the map}

\item{wd}{- an optional working directory (defaults to tempdir)}

\item{id}{- an optional id for the map - defaults to either the mapName or if not present the name of the zip file.}
}
\value{
a sf object containing the map
}
\description{
to standard code, name, altCode and codeType columns
}
\examples{
\dontrun{
downloadMap(
  zipUrl="https://opendata.arcgis.com/datasets/87511f3ae00a40208741c685f827c1d3_0.zip?outSR=\%7B\%22latestWkid\%22\%3A27700\%2C\%22wkid\%22\%3A27700\%7D",
  mapName="NHS_England_Regions__April_2020__Boundaries_EN_BGC",
  codeCol="nhser20cd",
  nameCol="nhser20nm"
)
}
}
