% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-maps.R
\name{downloadMap}
\alias{downloadMap}
\title{Download a shape file map}
\usage{
downloadMap(zipUrl, mapName = NULL, simplify = FALSE, ...)
}
\arguments{
\item{zipUrl}{the URL of the zipped shape file}

\item{mapName}{(optional) the layer name or map name - this is `"xyz"` if a
zip file contains `"xyz.shp"`. Usually there is one `.shp` file in a  zip
file, and by default it will be picked if this is not set.}

\item{simplify}{do you want to simplify the map?}

\item{...}{
  Named arguments passed on to \code{\link[=.cached]{.cached}}\describe{
    \item{\code{.nocache}}{an option to defeat the caching which can be set globally as options("cache.disable"=TRUE)}
\item{\code{.stale}}{the length of time in days to keep cached data before considering it as stale. can also be set by options("cache.stale")}
}}
}
\value{
a `sf` object containing the map
}
\description{
This function downloads and caches a zipped shape file map, unpacks it, finds
the correct shape file layer optionally simplifies it and converts it to `sf`.
}
\examples{
\dontrun{
downloadMap(
  zipUrl="https://bit.ly/3A9TnR1"
)
}
}
\concept{io}
