% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{getIntersection}
\alias{getIntersection}
\title{get the intersection between to maps with ids. Caches the result in the working directory.}
\usage{
getIntersection(inputShape, outputShape, ...)
}
\description{
get the intersection between to maps with ids. Caches the result in the working directory.
}
