% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{getIntersection}
\alias{getIntersection}
\title{Intersection of 2 shapes}
\usage{
getIntersection(
  inputShape,
  outputShape,
  suffix = c(".x", ".y"),
  recalcArea = TRUE,
  ...
)
}
\arguments{
\item{inputShape}{- the input sf}

\item{outputShape}{- the output sf}

\item{suffix}{- the suffix of any duplicated columns as per dplyr::inner_join()}

\item{recalcArea}{- do you need the area of the intersected shape (e.g. for areal interpolation)}

\item{...}{passed on to .cached() (cache control) - relevant is nocache = TRUE which prevents this from being precalculated}
}
\value{
a sf object representing the intersection of the input and output shapes.
}
\description{
get the intersection between to maps with ids. Caches the result in the working directory.
}
