% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-maps.R
\name{getMap}
\alias{getMap}
\title{Gets maps for which the metadata is known.}
\usage{
getMap(mapId, sources = .loadSources(...), codeType = mapId, ...)
}
\arguments{
\item{mapId}{- a name of a map}

\item{sources}{- a list of map sources - see `getOption("arear.mapsources",arear::mapsources)``}

\item{codeType}{- defaults to mapId, the codeType of the map}

\item{...}{- passed to .cache, param `nocache=TRUE` to disable caching}
}
\value{
a standard sf map
}
\description{
If a map needs to be downloaded as a shapefile then it is stored temporarily. The location of this download directory can be set as option("arear.download.dir" = "~/.)
}
\examples{
\dontrun{
map = getMap("NHSER20")
}
}
\concept{io}
