% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arear.R
\name{mapTheme}
\alias{mapTheme}
\title{A map theme to remove extraneous clutter}
\usage{
mapTheme()
}
\description{
A map theme to remove extraneous clutter
}
\examples{
ggplot2::ggplot()+mapTheme()
}
\concept{vis}
