% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-maps.R
\name{standardiseMap}
\alias{standardiseMap}
\title{Standardise maps to a minimal set of attributes with consistent naming with code, name and codeType columns and an optional altCode column}
\usage{
standardiseMap(sf, codeCol, nameCol, altCodeCol, codeType)
}
\arguments{
\item{sf}{- a non standard map}

\item{codeCol}{- a column name containing the unique code or id for the map}

\item{nameCol}{- the name column}

\item{altCodeCol}{- an alternative code column}

\item{codeType}{- a fixed value for the codeType}
}
\value{
a standardised map
}
\description{
Standardise maps to a minimal set of attributes with consistent naming with code, name and codeType columns and an optional altCode column
}
