% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surgecapacity}
\alias{surgecapacity}
\title{Hospital surge capacity in COVID}
\format{
A sf geometry with:
\describe{
  \item{nation}{England, Wales, etc...}
  \item{hospitalId}{An id for the hospital}
  \item{sector}{NHS or independent}
  \item{hospitalName}{the hospital name}
  \item{pcds}{the UK postcode of the hospital}
  \item{trustId}{the NHS trust or local health board of the hospital}
  \item{trustName}{the NHS trust or local health board name}
  \item{tier1}{indicator of the role of the hospital as an acure provider}
  \item{hduBeds}{the number of hdu beds the hospital could have provided at maximum surge in March 2020}
  \item{acuteBeds}{the number of acute beds the hospital could have provided at maximum surge in March 2020}
}
}
\usage{
surgecapacity
}
\description{
Locations of UK general medical hospitals in mid march 2020 with estimates of
beds available and maximal surge capacity HDU beds. This was manually
assembled and curated from various sources in mid march 2020 as the NHS
geared up to provide additional capacity to cope with the surge in COVID
cases. It is not an up to date picture of NHS capacity. It does not include
mental health or community hospitals. The surge capacity seems to have been
calculated quite differently in Scotland.
}
\concept{datasets}
\keyword{datasets}
