% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ukcovidmap}
\alias{ukcovidmap}
\title{Shape file related to the detailed level for PHE coronavirus statistical reporting in England, Wales, Scotland, and Northern Ireland on the coronavirus.gov.uk tracker.
The legacy is the level at which detailed data was reported in the first wave, prior to the API and was aggregated by https://github.com/tomwhite/covid-19-uk-data}
\source{
\describe{
  \item{England, Wales, Scotland & NI - LAD19}{https://geoportal.statistics.gov.uk/datasets/local-authority-districts-december-2019-boundaries-uk-buc}
  \item{England & Wales - LAD19}{https://geoportal.statistics.gov.uk/datasets/local-authority-districts-december-2019-boundaries-uk-buc}
  \item{Scotland - CA19}{https://www.spatialdata.gov.scot/geonetwork/srv/eng/catalog.search#/metadata/1cd57ea6-8d6e-412b-a9dd-d1c89a80ad62}
  \item{Northern Ireland  -LGD12}{https://data.gov.uk/dataset/05f72866-b72b-476a-b6f3-57bd4a768674/osni-open-data-largescale-boundaries-local-government-districts-2012}
}
}
\usage{
ukcovidmap(legacy = FALSE)
}
\arguments{
\item{legacy}{use the legacy version of the map from pre coronavirus api times.}
}
\value{
A 382 line data frame (or a 335 line data frame for legacy):
\describe{
  \item{code}{the ONS code for the ares}
  \item{name}{The name for the area}
  \item{codeType}{the type of area (LAD19)}
  \item{altCode}{NA}
}
}
\description{
This map matches the data reported by the PHE coronavirus api when it is downloading lower tier local authority regions e.g. via
https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&metric=newCasesBySpecimenDate&format=csv
There are 2 regions here for which no data is reported in this API - the city of London & the isles of scilly.
}
\details{
Data available under open government licence: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
}
